requirements.txt
Arpeggio==2.0.2
attrs==23.1.0
dill==0.3.7
grpcio==1.56.2
parver==0.5
protobuf==4.24.4
pulumi==3.88.1
pulumi-aws==6.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
semver==2.13.0
six==1.16.0